proyecto de validacion de formularios en thymeleaf

+++ com.jaimebackend.primeraweb/src/main/java/com/jaimebackend/primeraweb/service/PersonaService.java
@@ -16,5 +16,5 @@ public interface PersonaService {
 
     void eliminarPersona(Long id);
 
-    long contarPersonas();
+    int contarPersonas();
 }
