+++ com.jaimebackend.primeraweb/src/main/java/com/jaimebackend/primeraweb/service/impl/PersonaServiceImpl.java
@@ -46,8 +46,8 @@ private PersonaRepository personaRepository;
     }
 
     @Override
-    public long contarPersonas() {
-        return personaRepository.count();
+    public int contarPersonas() {
+        return Math.toIntExact(personaRepository.count());
     }
